Move transitions to more proper places
