+++ dnsconfd/fsm/transitions/stopping.py
@@ -1,15 +1,69 @@
 from dnsconfd.fsm import ContextEvent, ExitCode, ContextState
-from dnsconfd.fsm.transitions.with_exit import WithExit
+from dnsconfd.fsm.transitions import TransitionImplementations
 
 
-class Stopping(WithExit):
+class Stopping(TransitionImplementations):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.transitions = {
+            ContextState.STARTING: {
+                "STOP": (ContextState.STOPPING,
+                         self._exit_transition)
+            },
+            ContextState.CONFIGURING_DNS_MANAGER: {
+                "FAIL": (ContextState.STOPPING,
+                         self._exit_transition)
+            },
+            ContextState.CONNECTING_DBUS: {
+                "FAIL": (ContextState.STOPPING,
+                         self._exit_transition)
+            },
+            ContextState.SUBMITTING_START_JOB: {
+                "FAIL": (ContextState.STOPPING,
+                         self._exit_transition)
+            },
             ContextState.STOPPING: {
                 "EXIT": (ContextState.STOPPING,
                          self._exit_transition)
+            },
+            ContextState.WAITING_FOR_START_JOB: {
+                "START_FAIL": (ContextState.STOPPING,
+                               self._service_failure_exit_transition),
+                "STOP": (ContextState.WAITING_TO_SUBMIT_STOP_JOB,
+                         lambda y: None)
+            },
+            ContextState.POLLING: {
+                "TIMEOUT": (ContextState.STOPPING,
+                            self._exit_transition),
+                "STOP": (ContextState.REVERTING_RESOLV_CONF,
+                         self._running_stop_transition)
+            },
+            ContextState.UPDATING_RESOLV_CONF: {
+                "FAIL": (ContextState.SUBMITTING_STOP_JOB,
+                         self._updating_resolv_conf_fail_transition)
+            },
+            ContextState.UPDATING_ROUTES: {
+                "FAIL": (ContextState.REVERTING_RESOLV_CONF,
+                         self._running_stop_transition)
+            },
+            ContextState.UPDATING_DNS_MANAGER: {
+                "FAIL": (ContextState.REVERTING_RESOLV_CONF,
+                         self.updating_dns_manager_fail_transition)
+            },
+            ContextState.RUNNING: {
+                "STOP": (ContextState.REVERTING_RESOLV_CONF,
+                         self._running_stop_transition)
+            },
+            ContextState.SUBMITTING_RESTART_JOB: {
+                "FAIL": (ContextState.REVERT_RESOLV_ON_FAILED_RESTART,
+                         self._submitting_restart_job_fail_transition)
+            },
+            ContextState.WAITING_RESTART_JOB: {
+                "RESTART_FAIL": (ContextState.REVERT_RESOLV_ON_FAILED_RESTART,
+                                 self._waiting_restart_job_failure_transition),
+                "STOP": (ContextState.WAITING_TO_SUBMIT_STOP_JOB,
+                         lambda y: None)
             },
             ContextState.WAITING_TO_SUBMIT_STOP_JOB: {
                 "START_OK": (ContextState.SUBMITTING_STOP_JOB,
@@ -55,6 +109,13 @@ class Stopping(WithExit):
                 "FAIL": (ContextState.STOPPING, self._exit_transition)
             }
         }
+
+    def _service_failure_exit_transition(self, event: ContextEvent) \
+            -> ContextEvent | None:
+        self.container.set_exit_code(ExitCode.SERVICE_FAILURE)
+        self.lgr.info("Stopping event loop and FSM")
+        self.container.main_loop.quit()
+        return None
 
     def _to_removing_routes_transition(self, event: ContextEvent) \
             -> ContextEvent | None:
@@ -153,3 +214,97 @@ class Stopping(WithExit):
             ContextEvent("STOP_SUCCESS"),
             ContextEvent("STOP_FAILURE"))
         return ContextEvent("SUCCESS")
+
+    def _updating_resolv_conf_fail_transition(self, event: ContextEvent) \
+            -> ContextEvent | None:
+        """ Transition to SUBMITTING_STOP_JOB
+
+        Attempt to submit stop job
+
+        :param event: Not used
+        :type event: ContextEvent
+        :return: SUCCESS or FAIL with exit code
+        :rtype: ContextEvent | None
+        """
+        # since we have already problems with resolv.conf,
+        # we will be performing this without checking result
+        self.container.set_exit_code(ExitCode.RESOLV_CONF_FAILURE)
+        self.container.sys_mgr.revert_resolvconf()
+        if not self.container.subscribe_systemd_signals():
+            self.container.set_exit_code(ExitCode.DBUS_FAILURE)
+            return ContextEvent("FAIL")
+        service_stop_job = self.container.stop_unit()
+        if service_stop_job is None:
+            self.container.set_exit_code(ExitCode.DBUS_FAILURE)
+            return ContextEvent("FAIL")
+        self.container.systemd_jobs[service_stop_job] = (
+            ContextEvent("STOP_SUCCESS"),
+            ContextEvent("STOP_FAILURE"))
+        return ContextEvent("SUCCESS")
+
+    def updating_dns_manager_fail_transition(self, event: ContextEvent) \
+            -> ContextEvent | None:
+        """ Transition to REVERTING_RESOLV_CONF
+
+        Attempt to revert resolv.conf content
+
+        :param event: Not used
+        :type event: ContextEvent
+        :return: SUCCESS or FAIL with exit code
+        :rtype: ContextEvent | None
+        """
+        self.lgr.info("Failed to update DNS service, stopping")
+        if not self.container.sys_mgr.revert_resolvconf():
+            self.container.set_exit_code(ExitCode.RESOLV_CONF_FAILURE)
+            return ContextEvent("FAIL")
+        return ContextEvent("SUCCESS")
+
+    def _submitting_restart_job_fail_transition(self, event: ContextEvent) \
+            -> ContextEvent | None:
+        """ Transition to REVERTING_RESOLV_ON_FAILED_RESTART
+
+        Attempt to revert resolv.conf
+
+        :param event: Event with exit code in data
+        :type event: ContextEvent
+        :return: SUCCESS or FAIL with exit code
+        :rtype: ContextEvent | None
+        """
+        if not self.container.sys_mgr.revert_resolvconf():
+            self.lgr.error("Failed to revert resolv.conf")
+            self.container.set_exit_code(ExitCode.RESOLV_CONF_FAILURE)
+            return ContextEvent("FAIL")
+        self.lgr.debug("Successfully reverted resolv.conf")
+        return ContextEvent("SUCCESS", event.data)
+
+    def _waiting_restart_job_failure_transition(self, event: ContextEvent) \
+            -> ContextEvent | None:
+        """ Transition to REVERTING_RESOLV_ON_FAILED_RESTART
+
+        Attempt to revert resolv.conf
+
+        :param event: Not used
+        :type event: ContextEvent
+        :return: SUCCESS or FAIL with exit code
+        :rtype: ContextEvent | None
+        """
+        self.container.set_exit_code(ExitCode.SERVICE_FAILURE)
+        if not self.container.sys_mgr.revert_resolvconf():
+            self.lgr.error("Failed to revert resolv.conf")
+            return ContextEvent("FAIL")
+        self.lgr.debug("Successfully reverted resolv.conf")
+        return ContextEvent("SUCCESS")
+
+    def _exit_transition(self, event: ContextEvent):
+        """ Transition to STOPPING
+
+        Save exit code and stop main loop
+
+        :param event: Event holding exit code in data
+        :type event: ContextEvent
+        :return: None
+        :rtype: ContextEvent | None
+        """
+        self.lgr.info("Stopping event loop and FSM")
+        self.container.main_loop.quit()
+        return None
