+++ docs/generate_fsm.py
@@ -10,7 +10,7 @@ if __name__ == "__main__":
                            "resolv_conf_path": "/", "resolver_options": "edns0", "dnssec_enabled": True, "prioritize_wire": True, "handle_routing": True, "static_servers": {}}, None)
 
     g = graphviz.Digraph('G', filename='fsm.gv')
-    for (key, transitions) in ctx.transition.items():
+    for (key, transitions) in ctx.transitions.items():
         for (event, (next_state, cb)) in transitions.items():
             g.edge(key.name, next_state.name, label=event)
